bug: fixed issue with database seeding
