+++ script/seed.js
@@ -1,44 +1,74 @@
-'use strict'
+const {db, Product, User, Cart, OrderDetail} = require('../server/db/models')
+const users = require('../listOfUsers')
 
-const db = require('../server/db')
-const {User} = require('../server/db/models')
+const products = [
+  {
+    name: 'smash ball',
+    description: 'unlocks final smash',
+    imageUrl: 'https://www.ssbwiki.com/images/a/a2/SSBU_spirit_Smash_Ball.png',
+    price: 500,
+    rating: 5.0
+  },
+  {
+    name: 'blast box',
+    description: 'send your enemies flying, but be careful!',
+    imageUrl:
+      'https://www.ssbwiki.com/images/thumb/4/46/BlastBox.jpg/250px-BlastBox.jpg',
+    price: 250,
+    rating: 2.8
+  },
+  {
+    name: 'maxim tomato',
+    description: 'get your daily requirement of veggies',
+    imageUrl:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRYqDhfz2N1nKR9E9Lt4YbquzPmZNOtguUOdA&usqp=CAU',
+    price: 300,
+    rating: 3.5
+  },
+  {
+    name: 'bunny hood',
+    description: 'move faster than the wild!',
+    imageUrl:
+      'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSKzpkFIgj1pPrzT91FyG9xS1XfqDeL2VzekA&usqp=CAU',
+    price: 250,
+    rating: 4.2
+  }
+]
 
-async function seed() {
-  await db.sync({force: true})
-  console.log('db synced!')
-
-  const users = await Promise.all([
-    User.create({email: 'cody@email.com', password: '123'}),
-    User.create({email: 'murphy@email.com', password: '123'})
-  ])
-
-  console.log(`seeded ${users.length} users`)
-  console.log(`seeded successfully`)
-}
-
-// We've separated the `seed` function from the `runSeed` function.
-// This way we can isolate the error handling and exit trapping.
-// The `seed` function is concerned only with modifying the database.
-async function runSeed() {
-  console.log('seeding...')
+const seed = async () => {
   try {
-    await seed()
+    await db.sync({force: true})
+
+    await Promise.all(
+      products.map(product => {
+        return Product.create(product)
+      })
+    )
+
+    await Promise.all(
+      users.map(user => {
+        return User.create(user)
+      })
+    )
+
+    console.log('seeding done!')
+    // seed your database here!
   } catch (err) {
-    console.error(err)
-    process.exitCode = 1
-  } finally {
-    console.log('closing db connection')
-    await db.close()
-    console.log('db connection closed')
+    console.log(err)
   }
 }
 
-// Execute the `seed` function, IF we ran this module directly (`node seed`).
-// `Async` functions always return a promise, so we can use `catch` to handle
-// any errors that might occur inside of `seed`.
-if (module === require.main) {
-  runSeed()
-}
-
-// we export the seed function for testing purposes (see `./seed.spec.js`)
 module.exports = seed
+
+if (require.main === module) {
+  seed()
+    .then(() => {
+      console.log('Seeding success!')
+      db.close()
+    })
+    .catch(err => {
+      console.error('Oh noes! Something went wrong!')
+      console.error(err)
+      db.close()
+    })
+}
