+++ script/seed.spec.js
@@ -1,8 +1,8 @@
-'use strict'
-/* global describe it */
+// 'use strict'
+// /* global describe it */
 
-const seed = require('./seed')
+// const seed = require('./seed')
 
-describe('seed script', () => {
-  it('completes successfully', seed)
-})
+// describe('seed script', () => {
+//   it('completes successfully', seed)
+// })
