+++ server/db/models/user.spec.js
@@ -1,32 +1,32 @@
-/* global describe beforeEach it */
+// /* global describe beforeEach it */
 
-const {expect} = require('chai')
-const db = require('../index')
-const User = db.model('user')
+// const {expect} = require('chai')
+// const db = require('../index')
+// const User = db.model('user')
 
-describe('User model', () => {
-  beforeEach(() => {
-    return db.sync({force: true})
-  })
+// describe('User model', () => {
+//   beforeEach(() => {
+//     return db.sync({force: true})
+//   })
 
-  describe('instanceMethods', () => {
-    describe('correctPassword', () => {
-      let cody
+//   describe('instanceMethods', () => {
+//     describe('correctPassword', () => {
+//       let cody
 
-      beforeEach(async () => {
-        cody = await User.create({
-          email: 'cody@puppybook.com',
-          password: 'bones'
-        })
-      })
+//       beforeEach(async () => {
+//         cody = await User.create({
+//           email: 'cody@puppybook.com',
+//           password: 'bones'
+//         })
+//       })
 
-      it('returns true if the password is correct', () => {
-        expect(cody.correctPassword('bones')).to.be.equal(true)
-      })
+//       it('returns true if the password is correct', () => {
+//         expect(cody.correctPassword('bones')).to.be.equal(true)
+//       })
 
-      it('returns false if the password is incorrect', () => {
-        expect(cody.correctPassword('bonez')).to.be.equal(false)
-      })
-    }) // end describe('correctPassword')
-  }) // end describe('instanceMethods')
-}) // end describe('User model')
+//       it('returns false if the password is incorrect', () => {
+//         expect(cody.correctPassword('bonez')).to.be.equal(false)
+//       })
+//     }) // end describe('correctPassword')
+//   }) // end describe('instanceMethods')
+// }) // end describe('User model')
